requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-solo==2.1.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.4
